refactor GraphQL entities and DTOs to improve nullable handling and add ID type support

+++ src/modules/reviews/dto/update-review.dto.ts
@@ -1,9 +1,9 @@
-import { Field, Float, ArgsType } from '@nestjs/graphql';
+import { Field, ArgsType, Int } from '@nestjs/graphql';
 import { IsOptional, IsNumber, IsString } from 'class-validator';
 
 @ArgsType()
 export class UpdateReviewDto {
-  @Field(() => Float, { nullable: true })
+  @Field(() => Int, { nullable: true })
   @IsOptional()
   @IsNumber()
   rating?: number;
